% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{as.DTG.POSIXct}
\alias{as.DTG.POSIXct}
\title{Convert POSIXct class object to DTG class object}
\usage{
\method{as.DTG}{POSIXct}(x)
}
\arguments{
\item{x}{A POSIXct object.}
}
\description{
Convert POSIXct class object to DTG class object
}
\examples{
x <- DTG(c(a="220815ZDEC17", b="171900QJAN18"))
y <- as.POSIXct(x, tz="GMT")
z <- as.DTG(y)
}
