% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{as.POSIXct.DTG}
\alias{as.POSIXct.DTG}
\title{Convert DTG class object to POSIXct class object}
\usage{
\method{as.POSIXct}{DTG}(x)
}
\arguments{
\item{x}{A DTG object.}
}
\description{
Convert DTG class object to POSIXct class object
}
\examples{
x <- DTG("220815ZDEC17")
as.POSIXct(x)
}
