% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\docType{class}
\name{cgOpsum.list-class}
\alias{cgOpsum.list-class}
\alias{makeopsum.list}
\title{An S4 class to represent a cgOpsum list}
\description{
An S4 class to represent a cgOpsum list
}
\keyword{internal}
