% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{cgOpsum}
\alias{cgOpsum}
\title{Convert character vector of opsums to cgOpsum class}
\usage{
cgOpsum(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A list of class cgOpsum
}
\description{
Convert character vector of opsums to cgOpsum class
}
\examples{
file <- system.file("extdata", "07JAN15 OPSUM.txt", package = "cgToolkit")
x <- readtext::readtext(file)
opsum <- cgOpsum(x$text)
}
