% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{convertToDf}
\alias{convertToDf}
\title{Convert Text Strings Vector written in a Table Format to a new Data Frame}
\usage{
convertToDf(x, colname, cols = 2, idcolname, mynames = NULL)
}
\arguments{
\item{x}{A dataframe containing vectors of text strings and an index vector}

\item{colname}{Name of column containing table-like text strings}

\item{cols}{Integer representing number of columns in table, default value is 2}

\item{idcolname}{Name of column containing indices, usually doc_id if converted from a corpus}

\item{mynames}{Vector of column names for converted tabular data to pass to textTableToList function}
}
\description{
Convert Text Strings Vector written in a Table Format to a new Data Frame
}
