% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumExtractorFunctions.R
\name{data.frame.cgOpsum.list}
\alias{data.frame.cgOpsum.list}
\title{data frame Method for cgOpsum list}
\usage{
data.frame.cgOpsum.list(object)
}
\arguments{
\item{object}{A cgOpsum.list object.}
}
\description{
data frame Method for cgOpsum list
}
