% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{DTG}
\alias{DTG}
\title{Change character vector to a DTG vector.}
\usage{
DTG(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A DTG object
}
\description{
Change character vector to a DTG vector.
}
\examples{
DTG("220815ZDEC17")
DTG(c(dtg1 = "181918BDEC17", dtg2 = "182318BDEC17"))
\dontrun{
DTG("182918BDEC17") # Returns Warning
}

}
