% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumExtractorFunctions.R
\name{extractFuncs}
\alias{extractFuncs}
\title{Default Function to produce function map a  to .}
\usage{
extractFuncs(x = NULL)
}
\arguments{
\item{x}{A Named character vector.}
}
\value{
Named character vector
}
\description{
Default Function to produce function map a  to .
}
