% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{extractParagraph}
\alias{extractParagraph}
\title{Extract a paragraph of text from a corpus using start and end keywords.}
\usage{
extractParagraph(x, start, stop)
}
\arguments{
\item{x}{A corpus}

\item{start}{A character vector length 1 for the section start keyword}

\item{stop}{A character vector length 1 for the section end keyword}
}
\value{
A dataframe with doc_id and text columns
}
\description{
Extract a paragraph of text from a corpus using start and end keywords.
}
