% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{getLabeledData}
\alias{getLabeledData}
\title{Extract List of labeled text from a Corpus}
\usage{
getLabeledData(mycorpus, label)
}
\arguments{
\item{mycorpus}{A corpus object}

\item{label}{A regular expression to pass to grep

The function searches for data that matches the search pattern given by
label after splitting the corpus by lines of text.}
}
\description{
Extract List of labeled text from a Corpus
}
\examples{
file <- system.file("extdata", package = "opsumToolbox")
x <- readtext::readtext(file)
opsum <- quanteda::corpus(x$text)
quanteda::docnames(opsum) <- x$doc_id
y <- getLabeledData(opsum, label = "SUBJ:")
}
