% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{getStartStopPairs}
\alias{getStartStopPairs}
\title{Identify available start and stop paragraph heading keywords from a corpus}
\usage{
getStartStopPairs(x)
}
\arguments{
\item{x}{A corpus.}
}
\value{
A character vector of paragraph names
}
\description{
Identify available start and stop paragraph heading keywords from a corpus
}
