% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumExtractorFunctions.R
\name{getWeather}
\alias{getWeather}
\title{Default Function to convert a  to .}
\usage{
getWeather(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Named character vector
}
\description{
Default Function to convert a  to .
}
\keyword{internal}
