% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{mergeData}
\alias{mergeData}
\title{Create a text dataframe from a corpus using multiple start and end keywords.}
\usage{
mergeData(mycorpus, pair)
}
\arguments{
\item{mycorpus}{A corpus}

\item{pair}{A dataframe generated using the getStartStopPairs function}
}
\value{
A dataframe
}
\description{
Create a text dataframe from a corpus using multiple start and end keywords.
}
