% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{print.cgOpsum.list}
\alias{print.cgOpsum.list}
\title{print Method for cgOpsum}
\usage{
\method{print}{cgOpsum.list}(x, ...)
}
\arguments{
\item{x}{A cgOpsum object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print Method for cgOpsum
}
