% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\docType{methods}
\name{show,cgOpsum-method}
\alias{show,cgOpsum-method}
\title{Show Method for cgOpsum}
\usage{
\S4method{show}{cgOpsum}(object)
}
\arguments{
\item{object}{A cgOpsum object.}
}
\description{
Show Method for cgOpsum
}
