% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\docType{methods}
\name{show,cgOpsum.list-method}
\alias{show,cgOpsum.list-method}
\title{Show Method for cgOpsum.list}
\usage{
\S4method{show}{cgOpsum.list}(object)
}
\arguments{
\item{object}{A cgOpsum.list object.}
}
\description{
Show Method for cgOpsum.list
}
