% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\docType{methods}
\name{show,DTG-method}
\alias{show,DTG-method}
\title{Show Method for DTG}
\usage{
\S4method{show}{DTG}(object)
}
\arguments{
\item{object}{A DTG object.}
}
\description{
Show Method for DTG
}
