% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{show.cgOpsum}
\alias{show.cgOpsum}
\title{show Method for cgOpsum}
\usage{
show.cgOpsum(obj, i = 1L)
}
\arguments{
\item{obj}{A cgOpsum object}

\item{i}{An integer vector indicating which opsum to display}
}
\description{
show Method for cgOpsum
}
