% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{slice.cgOpsum.list}
\alias{slice.cgOpsum.list}
\title{slice Method for cgOpsum.list}
\usage{
\method{slice}{cgOpsum.list}(object, slice)
}
\arguments{
\item{object}{A cgOpsum.list object.}

\item{slice}{A vector of names or indices to subset}
}
\value{
A list
}
\description{
slice Method for cgOpsum.list
}
