% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{splitDTG}
\alias{splitDTG}
\title{Change character string DTG to an intermediate object.}
\usage{
splitDTG(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A list with values and DTG character representation
}
\description{
Change character string DTG to an intermediate object.
}
\keyword{internal}
