% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{splitParagraph}
\alias{splitParagraph}
\title{Converts a character textfile to semi-structured list by numbered paragraphs.}
\usage{
splitParagraph(dat)
}
\arguments{
\item{dat}{A character vector.}
}
\value{
A list of class cgOpsum
}
\description{
Converts a character textfile to semi-structured list by numbered paragraphs.
}
\keyword{internal}
