% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{[.cgOpsum.list}
\alias{[.cgOpsum.list}
\title{subset Method for cgOpsum.list}
\usage{
\method{[}{cgOpsum.list}(x, subset, ...)
}
\arguments{
\item{x}{A cgOpsum.list object.}

\item{subset}{A vector of names or indices to subset.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A cgOpsum.list object
}
\description{
subset Method for cgOpsum.list
}
