% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{subset.cgOpsum.list}
\alias{subset.cgOpsum.list}
\title{subset Method for cgOpsum.list}
\usage{
subset.cgOpsum.list(object, subset)
}
\arguments{
\item{object}{A cgOpsum.list object.}

\item{subset}{A vector of names or indices to subset}
}
\value{
A cgOpsum.list object
}
\description{
subset Method for cgOpsum.list
}
