% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgOpsumFunctionMethods.R
\name{textTableToList}
\alias{textTableToList}
\title{Convert Text String Vector stored in a Table Format to a list}
\usage{
textTableToList(x, cols = 2, id = NULL, mynames = NULL)
}
\arguments{
\item{x}{text string containing table like data for conversion using read.table}

\item{cols}{Integer representing number of columns in table, default value is 2}

\item{id}{vector of column indices, usually x$doc_id if converted from a corpus}

\item{mynames}{Vector of column names for converted tabular data}
}
\description{
Convert Text String Vector stored in a Table Format to a list
}
