% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{validDTG}
\alias{validDTG}
\title{Checks validity of DTG object.}
\usage{
validDTG(x)
}
\arguments{
\item{x}{A suspected DTG object.}
}
\value{
A logical vector
}
\description{
Checks validity of DTG object.
}
\keyword{internal}
