% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTG-ClassFunctions.R
\name{DTG}
\alias{DTG}
\alias{DTG}
\title{An S4 class to represent a date-time-group (DTG)}
\usage{
DTG(x)

DTG(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A DTG object
}
\description{
An S4 class to represent a date-time-group (DTG)

Change character string to a DTG object.
}
\examples{
DTG("220815ZDEC17")
DTG("181918BDEC17")
\dontrun{
DTG("182918BDEC17") # Returns Error
}

}
\keyword{internal}
